+++ yt_xarray/utilities/_utilities.py
@@ -1,12 +1,76 @@
 import importlib
 import os.path
-from typing import List, Optional, Tuple
+from os import PathLike
 
 import numpy as np
+import numpy.typing as npt
 import xarray as xr
 import yt.config
 
 from yt_xarray.accessor._xr_to_yt import known_coord_aliases
+
+
+class _DimConstructor:
+    def __init__(
+        self,
+        dim_min: float,
+        dim_max: float,
+        dim_n: int,
+        dim_name: str,
+        stretched: bool = False,
+    ):
+        self.dim_min = dim_min
+        self.dim_max = dim_max
+        self.dim_n = dim_n
+        self.dim_name = dim_name
+        self.stretched = stretched
+
+    @property
+    def dim_n_2(self) -> int:
+        return self.dim_n // 2
+
+    @property
+    def values(self) -> npt.NDArray:
+        dim_vals = np.linspace(self.dim_min, self.dim_max, self.dim_n)
+        if self.stretched:
+            dx = dim_vals[1:] - dim_vals[:-1]
+            dim_vals[self.dim_n_2 :] = dim_vals[self.dim_n_2 :] + dx[0] * 2.0
+        return dim_vals
+
+
+class _DsConstructor:
+    def __init__(
+        self,
+        min_x: float,
+        max_x: float,
+        n_x: int,
+        x_name: str,
+        min_y: float,
+        max_y: float,
+        n_y: int,
+        y_name: str,
+        min_z: float,
+        max_z: float,
+        n_z: int,
+        z_name: str,
+        z_units: str,
+        field_name: str,
+        coord_order: tuple[str, str, str],
+        is_stretched: dict[str, bool],
+    ):
+        self.x = _DimConstructor(
+            min_x, max_x, n_x, x_name, stretched=is_stretched[x_name]
+        )
+        self.y = _DimConstructor(
+            min_y, max_y, n_y, y_name, stretched=is_stretched[y_name]
+        )
+        self.z = _DimConstructor(
+            min_z, max_z, n_z, z_name, stretched=is_stretched[z_name]
+        )
+        self.z_units = z_units
+        self.field_name = field_name
+        self.coord_order = coord_order
+        self.is_stretched = is_stretched
 
 
 def construct_minimal_ds(
@@ -28,20 +92,27 @@ def construct_minimal_ds(
     z_stretched: bool = False,
     field_name: str = "test_field",
     n_fields: int = 1,
-    coord_order: Optional[Tuple[str, str, str]] = None,
+    coord_order: tuple[str, str, str] | list[str] | None = None,
     dtype: str = "float64",
     npseed: bool = False,
 ) -> xr.Dataset:
-    if coord_order is None:
-        coord_order = ("z", "y", "x")
+
+    valid_coord_order: tuple[str, str, str] = ("z", "y", "x")
+    if isinstance(coord_order, list):
+        if len(coord_order) == 3:
+            valid_coord_order = tuple(coord_order)  # type: ignore[assignment]
+        else:
+            raise RuntimeError("coord_order length must be 3")
+    elif isinstance(coord_order, tuple):
+        valid_coord_order = coord_order
 
     dtype_to_use = getattr(np, dtype)
 
     # contruct and return a minimal xarray dataset to use in tests as needed
 
     # current known z types = ["height", "depth", "radius", "altitude"]
-
-    cdict = dict(
+    is_stretched = {x_name: x_stretched, y_name: y_stretched, z_name: z_stretched}
+    cdict = _DsConstructor(
         min_x=min_x,
         max_x=max_x,
         n_x=n_x,
@@ -56,28 +127,21 @@ def construct_minimal_ds(
         z_name=z_name,
         z_units=z_units,
         field_name=field_name,
-        coord_order=coord_order,
+        coord_order=valid_coord_order,
+        is_stretched=is_stretched,
     )
 
-    is_stretched = {x_name: x_stretched, y_name: y_stretched, z_name: z_stretched}
-
     coords = {}
-    var_shape = ()
-    coord_order_rn = ()
+    var_shape: tuple[int, ...] = ()
+    coord_order_rn: tuple[str, ...] = ()
     coord_arrays = []
-    for c in coord_order:
-        mn = cdict["min_" + c]
-        mx = cdict["max_" + c]
-        n = cdict["n_" + c]
-        cname = c + "_name"
-        cvals = np.linspace(mn, mx, n)
-        if is_stretched[cdict[cname]]:
-            dx = cvals[1:] - cvals[:-1]
-            cvals[int(n / 2) :] = cvals[int(n / 2) :] + dx[0] * 2.0
-        coords[cdict[cname]] = cvals.astype(dtype_to_use)
-        coord_arrays.append(coords[cdict[cname]])
-        coord_order_rn += (cdict[cname],)
-        var_shape += (n,)
+    for c in valid_coord_order:
+        current: _DimConstructor = getattr(cdict, c)
+        cvals = current.values
+        coords[current.dim_name] = cvals.astype(dtype_to_use)
+        coord_arrays.append(coords[current.dim_name])
+        coord_order_rn += (current.dim_name,)
+        var_shape += (current.dim_n,)
 
     if npseed:
         np.random.seed(0)
@@ -86,22 +150,22 @@ def construct_minimal_ds(
     if n_fields > 1:
         data_vars = {}
         for i in range(n_fields):
-            fname = cdict["field_name"] + f"_{i}"
+            fname = cdict.field_name + f"_{i}"
             da = xr.DataArray(vals, coords=coord_arrays, dims=coord_order_rn)
             data_vars[fname] = da
     else:
         data_vars = {
-            cdict["field_name"]: xr.DataArray(
+            cdict.field_name: xr.DataArray(
                 vals, coords=coord_arrays, dims=coord_order_rn
             )
         }
 
-    other_atts = {"geospatial_vertical_units": cdict["z_units"]}
+    other_atts = {"geospatial_vertical_units": cdict.z_units}
 
     return xr.Dataset(data_vars=data_vars, attrs=other_atts)
 
 
-def _test_time_coord(nt=5):
+def _test_time_coord(nt: int = 5) -> npt.NDArray:
     t0 = np.datetime64("2001-01-02").astype("datetime64[ns]")
     dt = np.timedelta64(1, "D")
     tvals = [t0]
@@ -111,7 +175,7 @@ def _test_time_coord(nt=5):
 
 
 def _get_test_coord(
-    cname, n, minv: Optional[float] = None, maxv: Optional[float] = None
+    cname, n, minv: float | None = None, maxv: float | None = None
 ) -> np.ndarray:
     if cname in known_coord_aliases:
         cname = known_coord_aliases[cname]
@@ -144,9 +208,9 @@ def _get_test_coord(
 def construct_ds_with_extra_dim(
     icoord: int,
     dim_name: str = "time",
-    ncoords: Optional[int] = None,
+    ncoords: int | None = None,
     nd_space: int = 3,
-    reverse_indices: Optional[List[int]] = None,
+    reverse_indices: list[int] | None = None,
 ):
     coord_configs = {
         0: (dim_name, "x", "y", "z"),
@@ -178,7 +242,7 @@ def construct_ds_with_extra_dim(
     return xr.Dataset(data_vars=data_vars)
 
 
-def _find_file(file):
+def _find_file(file: PathLike[str] | str) -> PathLike[str] | str:
     if os.path.isfile(file):
         return file
 
@@ -187,7 +251,7 @@ def _find_file(file):
     if ddir != default_val:
         possible_file = os.path.join(ddir, file)
         if os.path.isfile(possible_file):
-            return possible_file
+            return str(possible_file)
 
     raise OSError(
         f"Could not find {file} in current directory or in the yt search path ({ddir})"
@@ -203,7 +267,7 @@ def _validate_file(function):
 
 
 def _import_optional_dep(
-    name: str, package: Optional[str] = None, custom_message: Optional[str] = None
+    name: str, package: str | None = None, custom_message: str | None = None
 ):
     # wrapper of importlib.import_module
     # name, package get sent to importlib.import_module